+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- API
+  - WeatherCondition.STRONG_WIND
+  
+### FIXED
+- [BUG: 188] STRONG_WIND Weather condition  
 
 ## [0.10.6] - 2019-03-02
 ### Added
@@ -192,3 +198,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.9.4]: https://github.com/coreGreenberet/homematicip-rest-api/compare/0.9.3.3...0.9.4
 
 [BUG: 141]: https://github.com/coreGreenberet/homematicip-rest-api/issues/141
+[BUG: 188]: https://github.com/coreGreenberet/homematicip-rest-api/issues/188